Fix bug in computing machines and parse slack files

+++ src/main/kotlin/science/atlarge/wta/simulator/WTASim.kt
@@ -116,7 +116,7 @@ object WTASim {
                         BigDecimal.valueOf(traceEndTime - traceStartTime)
                                 .multiply(BigDecimal.valueOf(resourcesPerMachine[i].toLong())) // number of resources
                                 .multiply(BigDecimal.valueOf(cli.targetUtilization!!)) // target util
-                                .multiply(BigDecimal.valueOf(machineFractions[i].toLong())), // Fraction of this machine
+                                .multiply(BigDecimal.valueOf(machineFractions[i])), // Fraction of this machine
                         32, RoundingMode.CEILING).setScale(0, RoundingMode.CEILING).intValueExact()
 
                 repeat(numMachines) { j ->
